pyproject.toml
[tool.poetry]
name = "cosmic_crypto_webservice"
version = "0.1.0"
description = ""
authors = ["Yehya Usman <yehya.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
uvicorn = "^0.14.0"
redis = "^3.5.3"
hiredis = "^2.0.0"
loguru = "^0.5.3"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-binance = "^1.0.12"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.7b0"
better-exceptions = "^0.3.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


